refactor: Move activity-specific selectors to dedicated activity.ts file
- Create new app/selectors/activity.ts for activity-specific selectors
- Move selectLocalTransactions and dedupeTransactions from transactionController.ts to activity.ts
- Create corresponding activity.test.ts with tests for selectLocalTransactions
- Update imports in UnifiedTransactionsView to use new activity selector module
- Remove unused imports (selectEvmAddress, areAddressesEqual) from transactionController.ts

This change decouples controller-level selectors from activity and smart transaction specific logic, matching the pattern used in the extension and improving code organization.
