+++ app/selectors/transactionController.test.ts
@@ -4,8 +4,7 @@ import { TransactionType } from '@metamask/transaction-controller';
 import {
   selectTransactions,
   selectLastWithdrawTokenByType,
-  selectLocalTransactions,
-  selectNonReplacedTransactions,
+    selectNonReplacedTransactions,
   selectRequiredTransactionIds,
   selectRequiredTransactionHashes,
   selectRequiredTransactions,
@@ -178,60 +177,6 @@ describe('TransactionController Selectors', () => {
       } as unknown as RootState;
 
       expect(selectRequiredTransactions(state)).toStrictEqual([child]);
-    });
-  });
-
-  describe('selectLocalTransactions', () => {
-    it('filters required child transactions before nonce dedupe', () => {
-      const activeEvmAddress = '0x0000000000000000000000000000000000000001';
-      const state = {
-        engine: {
-          backgroundState: {
-            AccountsController: {
-              internalAccounts: {
-                selectedAccount: 'account-1',
-                accounts: {
-                  'account-1': {
-                    id: 'account-1',
-                    address: activeEvmAddress,
-                    type: 'eip155:eoa',
-                  },
-                },
-              },
-            },
-            TransactionController: {
-              transactions: [
-                {
-                  id: 'child',
-                  hash: '0xCHILD',
-                  chainId: '0x1',
-                  time: 200,
-                  txParams: {
-                    from: activeEvmAddress,
-                    nonce: '0x1',
-                  },
-                },
-                {
-                  id: 'parent',
-                  chainId: '0x1',
-                  requiredTransactionIds: ['child'],
-                  time: 100,
-                  type: TransactionType.predictDeposit,
-                  txParams: {
-                    from: activeEvmAddress,
-                    nonce: '0x1',
-                  },
-                },
-              ],
-            },
-          },
-        },
-        pendingSmartTransactionsForGroup: [],
-      } as unknown as RootState;
-
-      expect(selectLocalTransactions(state)).toStrictEqual([
-        expect.objectContaining({ id: 'parent' }),
-      ]);
     });
   });
